fix column resize + cleanup
