+++ packages/main/src/components/AnalyticalTable/VirtualTableBodyContainer.tsx
@@ -1,34 +1,80 @@
-import { StyleClassHelper, ThemingParameters } from '@ui5/webcomponents-react-base';
-import React, { useEffect, useRef, useState } from 'react';
-import { GlobalStyleClasses } from '../..';
+import { StyleClassHelper } from '@ui5/webcomponents-react-base';
+import React, { useCallback, useEffect, useRef, useState } from 'react';
+import { GlobalStyleClasses } from '@ui5/webcomponents-react/lib/GlobalStyleClasses';
 
 export const VirtualTableBodyContainer = (props) => {
-  const { tableBodyHeight, totalColumnsWidth, children, parentRef, onScroll, isScrolling, classes } = props;
+  const {
+    tableBodyHeight,
+    totalColumnsWidth,
+    children,
+    parentRef,
+    classes,
+    infiniteScroll,
+    infiniteScrollThreshold,
+    onLoadMore,
+    selectionMode,
+    rows,
+    internalRowHeight
+  } = props;
   const [isMounted, setIsMounted] = useState(false);
+
   useEffect(() => {
     if (parentRef.current) {
       setIsMounted(true);
     }
   }, [parentRef.current]);
 
-  //todo
-  const classNames = StyleClassHelper.of(classes.tbody, GlobalStyleClasses.sapScrollBar);
+  const classNames = StyleClassHelper.of(classes.tbody, GlobalStyleClasses.sapScrollBar).className;
+
+  const lastScrollTop = useRef(0);
+  const firedInfiniteLoadEvents = useRef(new Set());
+
+  //todo test
+  const onScroll = useCallback(
+    (event) => {
+      const scrollOffset = event.target.scrollTop;
+      const isScrollingDown = lastScrollTop.current < scrollOffset;
+      if (isScrollingDown && infiniteScroll) {
+        lastScrollTop.current = scrollOffset;
+        const currentTopRow = Math.floor(scrollOffset / internalRowHeight);
+        if (rows.length - currentTopRow < infiniteScrollThreshold) {
+          if (!firedInfiniteLoadEvents.current.has(rows.length)) {
+            onLoadMore({
+              detail: {
+                rowCount: rows.length
+              }
+            });
+          }
+          firedInfiniteLoadEvents.current.add(rows.length);
+        }
+      }
+    },
+    [
+      infiniteScroll,
+      infiniteScrollThreshold,
+      onLoadMore,
+      rows.length,
+      internalRowHeight,
+      firedInfiniteLoadEvents,
+      lastScrollTop
+    ]
+  );
 
   return (
-    <div
-      className={classes.tbody}
-      ref={parentRef}
-      //todo
-      // onScroll={onScroll}
-      style={{
-        position: 'relative',
-        overflowY: 'auto',
-        pointerEvents: isScrolling ? 'none' : 'auto',
-        height: `${tableBodyHeight}px`,
-        width: `${totalColumnsWidth}px`
-      }}
-    >
-      {isMounted && children}
-    </div>
+    <>
+      <div
+        className={classNames}
+        ref={parentRef}
+        onScroll={onScroll}
+        style={{
+          position: 'relative',
+          overflowY: 'auto',
+          height: `${tableBodyHeight}px`,
+          width: `${totalColumnsWidth}px`
+        }}
+      >
+        {isMounted && children}
+      </div>
+    </>
   );
 };
