+++ packages/main/src/components/AnalyticalTable/hooks/useStyling.ts
@@ -1,23 +1,10 @@
 import { getRTL } from '@ui5/webcomponents-base/dist/config/RTL';
-import { StyleClassHelper } from '@ui5/webcomponents-react-base/lib/StyleClassHelper';
 import { ThemingParameters } from '@ui5/webcomponents-react-base/lib/ThemingParameters';
-import { GlobalStyleClasses } from '@ui5/webcomponents-react/lib/GlobalStyleClasses';
 import { TableSelectionBehavior } from '@ui5/webcomponents-react/lib/TableSelectionBehavior';
 import { TableSelectionMode } from '@ui5/webcomponents-react/lib/TableSelectionMode';
 import { TextAlign } from '@ui5/webcomponents-react/lib/TextAlign';
 import { VerticalAlign } from '@ui5/webcomponents-react/lib/VerticalAlign';
 import { CSSProperties } from 'react';
-
-const getTableProps = (tableProps, { instance }) => {
-  const { classes } = instance.webComponentsReactProperties;
-  const classNames = StyleClassHelper.of(classes.table, GlobalStyleClasses.sapScrollBar).className;
-  return [
-    tableProps,
-    {
-      className: classNames
-    }
-  ];
-};
 
 const getHeaderGroupProps = (headerGroupProps, { instance }) => {
   const { classes } = instance.webComponentsReactProperties;
@@ -160,7 +147,6 @@ const getCellProps = (cellProps, { cell: { column }, instance }) => {
 };
 
 export const useStyling = (hooks) => {
-  hooks.getTableProps.push(getTableProps);
   hooks.getHeaderGroupProps.push(getHeaderGroupProps);
   hooks.getHeaderProps.push(getHeaderProps);
   hooks.getRowProps.push(getRowProps);
